requirements.txt
flask
flask_sqlalchemy
python-docx